add doc comment and move tests
